requirements.txt
django-graphos==0.3.41
xlrd==1.1.0
django-encrypted-model-fields==0.5.3